requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
celery
redis
sqlalchemy
python-jose[cryptography]
boto3
passlib
python-dotenv
